requirements.txt
packaging>=25.0
peewee>=3.18.3
pillow>=12.0.0
PySide6>=6.10
PySide6_Addons>=6.10
PySide6_Essentials>=6.10
pytesseract>=0.3.13
shiboken6>=6.10
tomli_w>=1.2.0
pynput>=1.8.0